chore(reclamation): size-cap extractions, full theme-token sweep, zero-warning build
Closes the 2026-06-28 reclamation backlog (A1, A2, B1, B2, B3) plus the
UCS mojibake known issue:

- MainComponent.h 2,132 -> 1,927 lines (back under its frozen 2,084
  ceiling): SelectionInfoPanel and CallbackDocumentWindow extracted to
  Source/UI/; status-bar and overlay chrome now theme-driven.
- RegionListPanel.cpp 1,533 -> 1,319 lines: batch-rename pipeline moved
  to RegionListPanel_BatchRename.cpp.
- Hardcoded-color audit (Sec 6.11): ~130 chrome literals across 18 files
  (plugin scan dialogs, plugin paths panel, Batch Processor, waveform
  selection accents, recording dialog, misc dialogs) replaced with
  ThemeManager tokens; long-lived components re-apply on theme switch;
  text on colored surfaces uses brightness-computed contrast.
- Clean-rebuild warning audit: all 184 non-deprecation warning sites
  fixed behavior-preservingly (explicit casts, juce::exactlyEqual for
  bit-exact change detection, unused-symbol removal, exhaustive switch
  cases, ctor-init reorder, shadow renames). Remaining: 39 JUCE 8 API
  deprecation sites, deferred as a dedicated migration (TODO.md).
- UCS keyword table regenerated with mojibake repair; generator now
  normalizes keywords to ASCII and fails on non-ASCII output.

Verified: 512 test groups / 207,782 assertions / 0 failures; zero
non-deprecation compiler warnings on a clean rebuild; ASCII, file-size,
and README-keymap gates all pass; adversarial code review found no
critical/high issues (its 4 findings are fixed in this commit).

+++ Source/DSP/TimePitchEngine.cpp
@@ -46,8 +46,8 @@ juce::AudioBuffer<float> apply(const juce::AudioBuffer<float>& input,
         c.reserve(static_cast<size_t>(numSamples) * 2);
 
     constexpr int kChunkFrames = 4096;
-    std::vector<float> interleavedIn (static_cast<size_t>(kChunkFrames) * numChannels);
-    std::vector<float> interleavedOut(static_cast<size_t>(kChunkFrames) * numChannels);
+    std::vector<float> interleavedIn (static_cast<size_t>(kChunkFrames) * static_cast<size_t>(numChannels));
+    std::vector<float> interleavedOut(static_cast<size_t>(kChunkFrames) * static_cast<size_t>(numChannels));
 
     auto drainOutput = [&]()
     {
